+++ CHANGELOG.md
@@ -3,7 +3,17 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - Unreleased
+## [1.4.0] - 2025/09/19
+
+### Added
+
+- The pnc transformation: leverages the categorize, earliest, and age transformations in sequence. The column names for this transformation are explicit and required. [PR #24](https://github.com/phac-nml/metadatatransformation/pull/24)
+
+### Changed
+
+- 16 columns of metadata are now accepted by all transformations. [PR #24](https://github.com/phac-nml/metadatatransformation/pull/24)
+
+## [1.3.0] - 2025/09/09
 
 ### Added
 
@@ -60,3 +70,4 @@ Initial release of phac-nml/metadatatransformation.
 [1.1.1]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.1.1
 [1.2.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.2.0
 [1.3.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.3.0
+[1.4.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.4.0